refactor: migrate registry imports from xulcan.system to xulcan.registry. close #22
