+++ app/xulcan/kernel/runtime.py
@@ -41,7 +41,7 @@ from xulcan.kernel.interfaces import (
     SentinelStrategy,
     HumanGateStrategy,
 )
-from xulcan.system.registry import ProviderRegistry
+from xulcan.registry import ProviderRegistry
 from xulcan.kernel.environment import SystemEnvironment
 
 from xulcan.blueprint.schema import AgentBlueprint
