+++ app/xulcan/registry/__init__.py
@@ -1,0 +1,9 @@
+"""Xulcan Registry Package.
+
+This package provides infrastructure for dependency injection and credential management.
+"""
+
+from .base import ProviderRegistry
+from .credentials import CredentialProxy, ToolSecretsVault
+
+__all__ = ["ProviderRegistry", "CredentialProxy", "ToolSecretsVault"]
